pyproject.toml
[project]
name = "checksqs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "azure-storage-blob>=12.25.1",
    "boto3>=1.38.12",
    "celery-prometheus-exporter>=1.7.0",
    "celery[mongodb,sqs]>=5.5.2",
    "firecrawl>=2.9.0",
    "flower>=2.0.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.167.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.4",
    "googleapis-common-protos==1.70.0",
    "load-dotenv>=0.1.0",
    "pillow>=11.2.1",
    "praw>=7.8.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "tavily-python>=0.7.7",
    "v>=1",
]
